// react-coding-problems/src/Projects/Trello-Board/models/models.tsx
export interface ITasks {
  boardId: number;
  boardName: string;
  tasks: ITask[];
}
export interface ITask {
  id: number;
  title: string;
  description: string;
  author: string;
  createdAt: string;
}
export interface IDraggedItem {
  item: ITask | null;
  fromBoardId: number | null;
  toBoardId: number | null;
}
export enum TaskClassNames {
  ongoing = "ongoing-tasks",
  todo = "todo-tasks",
  completed = "completed-tasks",
}